requirements.txt
# Project Dependencies

# Core
python-dateutil>=2.8.2

# Supabase (database connection)
supabase>=2.3.0
python-dotenv>=1.0.0

# FastAPI (API layer)
# fastapi>=0.104.0
# uvicorn>=0.24.0
# pydantic>=2.5.0

# Vector embeddings (local model, no API cost)
# sentence-transformers>=2.2.0

# Document parsing
PyPDF2>=3.0.0
python-docx>=1.0.0
python-pptx>=0.6.21
python-multipart>=0.0.6

# Moodle integration
beautifulsoup4>=4.12.0
httpx>=0.25.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
